+++ tap_xero/client.py
@@ -2,14 +2,107 @@ from base64 import b64encode
 import re
 import json
 import decimal
+import sys
+import math
 from os.path import join
 from datetime import datetime, date, time, timedelta
 import requests
 from singer.utils import strftime, strptime_to_utc
 import six
 import pytz
+import backoff
+import singer
+
+LOGGER = singer.get_logger()
 
 BASE_URL = "https://api.xero.com/api.xro/2.0"
+
+
+class XeroError(Exception):
+    def __init__(self, message=None, response=None):
+        super().__init__(message)
+        self.message = message
+        self.response = response
+
+
+class XeroBadRequestError(XeroError):
+    pass
+
+
+class XeroUnauthorizedError(XeroError):
+    pass
+
+
+class XeroForbiddenError(XeroError):
+    pass
+
+
+class XeroNotFoundError(XeroError):
+    pass
+
+
+class XeroPreConditionFailedError(XeroError):
+    pass
+
+
+class XeroTooManyError(XeroError):
+    pass
+
+
+class XeroTooManyInMinuteError(XeroError):
+    pass
+
+
+class XeroInternalError(XeroError):
+    pass
+
+
+class XeroNotImplementedError(XeroError):
+    pass
+
+
+class XeroNotAvailableError(XeroError):
+    pass
+
+
+ERROR_CODE_EXCEPTION_MAPPING = {
+    400: {
+        "raise_exception": XeroBadRequestError,
+        "message": "A validation exception has occurred."
+    },
+    401: {
+        "raise_exception": XeroUnauthorizedError,
+        "message": "Invalid authorization credentials."
+    },
+    403: {
+        "raise_exception": XeroForbiddenError,
+        "message": "User doesn't have permission to access the resource."
+    },
+    404: {
+        "raise_exception": XeroNotFoundError,
+        "message": "The resource you have specified cannot be found."
+    },
+    412: {
+        "raise_exception": XeroPreConditionFailedError,
+        "message": "One or more conditions given in the request header fields were invalid."
+    },
+    429: {
+        "raise_exception": XeroTooManyError,
+        "message": "The API rate limit for your organisation/application pairing has been exceeded"
+    },
+    500: {
+        "raise_exception": XeroInternalError,
+        "message": "An unhandled error with the Xero API. Contact the Xero API team if problems persist."
+    },
+    501: {
+        "raise_exception": XeroNotImplementedError,
+        "message": "The method you have called has not been implemented."
+    },
+    503: {
+        "raise_exception": XeroNotAvailableError,
+        "message": "API service is currently unavailable."
+    }
+}
 
 
 def parse_date(value):
@@ -66,6 +159,24 @@ def update_config_file(config, config_path):
     with open(config_path, 'w') as config_file:
         json.dump(config, config_file, indent=2)
 
+def is_not_status_code_fn(status_code):
+    def gen_fn(exc):
+        if getattr(exc, 'response', None) and getattr(exc.response, 'status_code', None) and exc.response.status_code not in status_code:
+            return True
+        # Retry other errors up to the max
+        return False
+    return gen_fn
+
+def retry_after_wait_gen():
+    while True:
+        # This is called in an except block so we can retrieve the exception
+        # and check it.
+        exc_info = sys.exc_info()
+        resp = exc_info[1].response
+        sleep_time_str = resp.headers.get('Retry-After')
+        LOGGER.info("API rate limit exceeded -- sleeping for %s seconds", sleep_time_str)
+        yield math.floor(float(sleep_time_str))
+
 class XeroClient():
     def __init__(self, config):
         self.session = requests.Session()
@@ -87,15 +198,43 @@ class XeroClient():
             "refresh_token": config["refresh_token"],
         }
         resp = self.session.post("https://identity.xero.com/connect/token", headers=headers, data=post_body)
-        resp.raise_for_status()
-        resp = resp.json()
 
-        # Write to config file
-        config['refresh_token'] = resp["refresh_token"]
-        update_config_file(config, config_path)
-        self.access_token = resp["access_token"]
-        self.tenant_id = config['tenant_id']
+        if resp.status_code != 200:
+            raise_for_error(resp)
+        else:
+            resp = resp.json()
 
+            # Write to config file
+            config['refresh_token'] = resp["refresh_token"]
+            update_config_file(config, config_path)
+            self.access_token = resp["access_token"]
+            self.tenant_id = config['tenant_id']
+
+
+    @backoff.on_exception(backoff.expo, (json.decoder.JSONDecodeError, XeroInternalError), max_tries=3)
+    @backoff.on_exception(retry_after_wait_gen, XeroTooManyInMinuteError, giveup=is_not_status_code_fn([429]), jitter=None, max_tries=3)
+    def check_platform_access(self, config, config_path):
+
+        # Validating the authentication of the provided configuration
+        self.refresh_credentials(config, config_path)
+
+        headers = {
+            "Authorization": "Bearer " + self.access_token,
+            "Xero-Tenant-Id": self.tenant_id,
+            "Content-Type": "application/json"
+        }
+
+        # Validating the authorization of the provided configuration
+        currencies_url = join(BASE_URL, "Currencies")
+        request = requests.Request("GET", currencies_url, headers=headers)
+        response = self.session.send(request.prepare())
+
+        if response.status_code != 200:
+            raise_for_error(response)
+
+
+    @backoff.on_exception(backoff.expo, (json.decoder.JSONDecodeError, XeroInternalError), max_tries=3)
+    @backoff.on_exception(retry_after_wait_gen, XeroTooManyInMinuteError, giveup=is_not_status_code_fn([429]), jitter=None, max_tries=3)
     def filter(self, tap_stream_id, since=None, **params):
         xero_resource_name = tap_stream_id.title().replace("_", "")
         url = join(BASE_URL, xero_resource_name)
@@ -109,9 +248,56 @@ class XeroClient():
 
         request = requests.Request("GET", url, headers=headers, params=params)
         response = self.session.send(request.prepare())
-        response.raise_for_status()
-        response_meta = json.loads(response.text,
-                                   object_hook=_json_load_object_hook,
-                                   parse_float=decimal.Decimal)
-        response_body = response_meta.pop(xero_resource_name)
-        return response_body
+
+        if response.status_code != 200:
+            raise_for_error(response)
+            return None
+        else:
+            response_meta = json.loads(response.text,
+                                    object_hook=_json_load_object_hook,
+                                    parse_float=decimal.Decimal)
+            response_body = response_meta.pop(xero_resource_name)
+            return response_body
+
+
+def raise_for_error(resp):
+    try:
+        resp.raise_for_status()
+    except (requests.HTTPError, requests.ConnectionError) as error:
+        try:
+            error_code = resp.status_code
+
+            # Handling status code 429 specially since the required information is present in the headers
+            if error_code == 429:
+                resp_headers = resp.headers
+                api_rate_limit_message = ERROR_CODE_EXCEPTION_MAPPING[429]["message"]
+                message = "HTTP-error-code: 429, Error: {}. Please retry after {} seconds".format(api_rate_limit_message, resp_headers.get("Retry-After"))
+
+                #Raise XeroTooManyInMinuteError exception if minute limit is reached
+                if resp_headers.get("X-Rate-Limit-Problem") == 'minute':
+                    raise XeroTooManyInMinuteError(message, resp) from None
+            # Handling status code 403 specially since response of API does not contain enough information
+            elif error_code in (403, 401):
+                api_message = ERROR_CODE_EXCEPTION_MAPPING[error_code]["message"]
+                message = "HTTP-error-code: {}, Error: {}".format(error_code, api_message)
+            else:
+                # Forming a response message for raising custom exception
+                try:
+                    response_json = resp.json()
+                except Exception:
+                    response_json = {}
+
+                message = "HTTP-error-code: {}, Error: {}".format(
+                    error_code,
+                    response_json.get(
+                        "error", response_json.get(
+                            "Title", response_json.get(
+                                "Detail", ERROR_CODE_EXCEPTION_MAPPING.get(
+                                    error_code, {}).get("message", "Unknown Error")
+                                ))))
+
+            exc = ERROR_CODE_EXCEPTION_MAPPING.get(error_code, {}).get("raise_exception", XeroError)
+            raise exc(message, resp) from None
+
+        except (ValueError, TypeError):
+            raise XeroError(error) from None